+++ run_comb_annot.sh
@@ -2,7 +2,7 @@
 set -euo pipefail
 
 comb_annot=$1
-BLAST=$2
+# BLAST=$2
 out_name=$3
 maker_dir=$4
 out_dir=$5
@@ -14,8 +14,10 @@ SCRIPTS=$7
 cd $comb_annot
 rm -rf $comb_annot/ref.*
 ln -s $PROTEIN1 $comb_annot/ref1_protein.fasta
-$SCRIPTS/combined_annot.sh $out_name $comb_annot $maker_dir/seq.fasta $out_dir/annot $maker_dir $snap_dir $SCRIPTS SGD 90 $comb_annot # results in $comb_annot/gff/$out_name.genes.gff
 makeblastdb -in ref1_protein.fasta -dbtype prot -parse_seqids -out ref
+if [ ! -e $comb_annot/gff/$out_name.genes.gff ]; then
+  $SCRIPTS/combined_annot.sh $out_name $comb_annot $maker_dir/seq.fasta $out_dir/annot $maker_dir $snap_dir $SCRIPTS SGD 90 $comb_annot # results in $comb_annot/gff/$out_name.genes.gff
+fi
 
 rm -rf $comb_annot/ref.*
 ln -s $PROTEIN2 $comb_annot/ref_protein.fasta
@@ -25,11 +27,15 @@ ln -s $CFG_DIR/maker_opts.ctl $comb_annot/maker_opts.ctl
 ln -s $CFG_DIR/maker_bopts.ctl $comb_annot/maker_bopts.ctl
 ln -s $CFG_DIR/maker_exe.ctl $comb_annot/maker_exe.ctl
 
-$SCRIPTS/unannot_regions.sh $maker_dir/seq.fasta $comb_annot/gff/$out_name.genes.gff $comb_annot $SCRIPTS # resutls in $comb_annot/non_orf.fasta
-rm $comb_annot/seq.fasta
 makeblastdb -in ref_protein.fasta -dbtype prot -parse_seqids -out ref
+if [ ! -e $comb_annot/non_orf.fasta ]; then
+  $SCRIPTS/unannot_regions.sh $maker_dir/seq.fasta $comb_annot/gff/$out_name.genes.gff $comb_annot $SCRIPTS # resutls in $comb_annot/non_orf.fasta
+fi
+if [ -e $comb_annot/seq.fasta ]; then rm $comb_annot/seq.fasta; fi
 ln -s $comb_annot/non_orf.fasta $comb_annot/seq.fasta
-$SCRIPTS/maker.sh $comb_annot $snap_dir $SCRIPTS # results in $comb_annot/genes.gff
+if [ ! -e $comb_annot/genes.gff ]; then
+  $SCRIPTS/maker.sh $comb_annot $snap_dir $SCRIPTS # results in $comb_annot/genes.gff
+fi
 less $comb_annot/non_orf.fasta | grep ">" > $comb_annot/head.txt
 mv $comb_annot/genes.gff $comb_annot/add1.genes.gff
 $BIN/conv_scf_pos $comb_annot/head.txt $comb_annot/add1.genes.gff > $comb_annot/genes.gff
@@ -37,7 +43,9 @@ $BIN/conv_scf_pos $comb_annot/head.txt $comb_annot/add1.genes.gff > $comb_annot/
 
 echo "#" > $comb_annot/$out_name.codex
 mkdir -p $comb_annot/more_annot
-$SCRIPTS/combined_annot.sh $out_name $comb_annot/more_annot $maker_dir/seq.fasta $comb_annot $comb_annot $snap_dir $SCRIPTS ENSEMBL 80 $comb_annot # results in $comb_annot/more_annot/gff/$out_name.genes.gff
+if [ ! -e $comb_annot/more_annot/gff/$out_name.genes.gff ]; then
+  $SCRIPTS/combined_annot.sh $out_name $comb_annot/more_annot $maker_dir/seq.fasta $comb_annot $comb_annot $snap_dir $SCRIPTS ENSEMBL 80 $comb_annot # results in $comb_annot/more_annot/gff/$out_name.genes.gff
+fi
 less $comb_annot/more_annot/gff/$out_name.genes.gff >> $comb_annot/gff/$out_name.genes.gff
 less $comb_annot/more_annot/blast_out/$out_name.blastx.out >> $comb_annot/blast_out/$out_name.blastx.out
 rm -rf $comb_annot/more_annot
@@ -45,7 +53,9 @@ rm -rf $comb_annot/more_annot
 rm -rf $comb_annot/seq.fasta
 rm -rf $comb_annot/non_orf.fasta
 rm -rf $comb_annot/genes.gff
-$SCRIPTS/unannot_regions.sh $maker_dir/seq.fasta $comb_annot/gff/$out_name.genes.gff $comb_annot $SCRIPTS # resutls in $comb_annot/non_orf.fasta
+if [ ! -e $comb_annot/non_orf.fasta ]; then
+  $SCRIPTS/unannot_regions.sh $maker_dir/seq.fasta $comb_annot/gff/$out_name.genes.gff $comb_annot $SCRIPTS # resutls in $comb_annot/non_orf.fasta
+fi
 
 augustus --gff3=on --species=$AUGUSTUS_REF non_orf.fasta > $comb_annot/genes.gff
 less $comb_annot/genes.gff | sed '/^#/d' | sed '/^$/d' | grep -E "gene|CDS" > $comb_annot/temp.gff
@@ -57,8 +67,10 @@ less $comb_annot/non_orf.fasta | grep ">" > $comb_annot/head.txt
 $BIN/conv_scf_pos $comb_annot/head.txt $comb_annot/add2.genes.gff > $comb_annot/genes.gff
 
 echo "#" > $comb_annot/$out_name.codex
-mkdir -p $comb_annot/more_annot
-$SCRIPTS/combined_annot.sh $out_name $comb_annot/more_annot $maker_dir/seq.fasta $comb_annot $comb_annot $snap_dir $SCRIPTS ENSEMBL 80 $comb_annot # results in $comb_annot/more_annot/gff/$out_name.genes.gff
+if [ ! -e $comb_annot/more_annot/gff/$out_name.genes.gff ]; then
+  mkdir -p $comb_annot/more_annot
+  $SCRIPTS/combined_annot.sh $out_name $comb_annot/more_annot $maker_dir/seq.fasta $comb_annot $comb_annot $snap_dir $SCRIPTS ENSEMBL 80 $comb_annot # results in $comb_annot/more_annot/gff/$out_name.genes.gff
+fi
 less $comb_annot/more_annot/gff/$out_name.genes.gff >> $comb_annot/gff/$out_name.genes.gff
 less $comb_annot/more_annot/blast_out/$out_name.blastx.out >> $comb_annot/blast_out/$out_name.blastx.out
 rm -rf $comb_annot/more_annot
